+++ src/Response/Types/AbstractType.php
@@ -6,7 +6,6 @@ use InvalidArgumentException;
 use League\Fractal\Resource\ResourceAbstract;
 use Somnambulist\Bundles\ApiBundle\Response\ResponseTypeInterface;
 use function array_is_list;
-use function trigger_deprecation;
 
 /**
  * Class ObjectResponse
@@ -24,30 +23,11 @@ abstract class AbstractType implements ResponseTypeInterface
 
     abstract public function asResource(): ResourceAbstract;
 
-    public function withKey(string $key): self
-    {
-        trigger_deprecation('somnambulist/api-bundle', '3.8.0', 'Use key() instead');
-
-        return $this->key($key);
-    }
-
     public function key(string $key): self
     {
         $this->key = $key;
 
         return $this;
-    }
-
-    public function withIncludes(...$includes): self
-    {
-        trigger_deprecation('somnambulist/api-bundle', '3.8.0', 'Use include() instead');
-
-        if (count($includes) === 1 && is_array($includes[0])) {
-            trigger_deprecation('somnambulist/api-bundle', '3.6.0', 'Passing an array of includes is deprecated. Use multiple string arguments');
-            $includes = $includes[0];
-        }
-
-        return $this->include(...$includes);
     }
 
     public function include(string ...$includes): self
@@ -72,13 +52,6 @@ abstract class AbstractType implements ResponseTypeInterface
         $this->fields = $fields;
 
         return $this;
-    }
-
-    public function withMeta(array $meta): self
-    {
-        trigger_deprecation('somnambulist/api-bundle', '3.8.0', 'Use meta() instead');
-
-        return $this->meta($meta);
     }
 
     public function meta(array $meta): self